+++ src/api/modules/sponsor/index.ts
@@ -1,0 +1,2 @@
+export { default as getEdition } from './getEdition'
+export { default as setEdition } from './setEdition'