+++ src/modules/apps/deploy.ts
@@ -7,7 +7,7 @@ import { parseLocator, toAppLocator } from '../../api/locator'
 import log from '../../api/logger'
 
 import { returnToPreviousAccount, switchAccount } from '../auth/switch'
-import { promptConfirm } from '../prompts'
+import { promptConfirm } from '../../api/modules/prompts'
 
 const switchToVendorMessage = (vendor: string): string => {
   return `You are trying to deploy this app in an account that differs from the indicated vendor. Do you want to deploy in account ${chalk.blue(