+++ src/modules/apps/undeprecate.ts
@@ -6,7 +6,7 @@ import { SessionManager } from '../../api/session/SessionManager'
 import { parseLocator } from '../../api/locator'
 import log from '../../api/logger'
 import { returnToPreviousAccount, switchAccount } from '../auth/switch'
-import { promptConfirm } from '../prompts'
+import { promptConfirm } from '../../api/modules/prompts'
 import { TelemetryCollector } from '../../lib/telemetry/TelemetryCollector'
 
 let originalAccount